Fix bdiff test failures on AIX.
The test fails when, in the call to calloc(), the number of elements (bn)
is 0. In that case, calloc() on AIX will return NULL, while the code expects
a valid heap pointer. Both results are permissible under C99, Unix98 etc.

Work around by ensuring that at least 1 element is requested.