Bundle python dll in win32 installer if present
in case hg.exe is generated with the py2exe '-b 2' option instead of '-b 1',
python2?.dll is not included in library.zip file and must also be packaged.
The file will be integrated in the installer only if it is present in the dist
directory.  Generating hg.exe with the '-b 2' option is mandatory to avoid a
crash when pyreadline is installed.