hgrc.5: reformatted to avoid big chunks of monospaced text
The man page used to have lines longer than 80 characters, even though
all lines in the hgrc.5.txt file were wrapped nicely. The problem
turned out to be that the indented paragraphs started literal blocks
instead of normal paragraphs. The literal blocks were of course not
wrapped when displayed by man.

In short, the asciidoc rules require lists to be formatted like this:

foo::
  Some description of foo.
+
Another paragraph in the description of foo. It *must* start flush
left and the plus is necessary to indicate that this is a list item
continuation.

Lists with nested lists can be formatted correctly using something
called "open blocks". These blocks are used to group the list items
and are marked by a line above and below with two dashes. See the
asciidoc user guide for the gory details...