setup: ignore 'not importing' warnings during version detection
Python will issue an ImportWarning when seeing 'import locale' if
there is a locale/ directory present without a __init__.py file.

The warning is silent by default, but it somehow shows up anyway on
Windows when setup.py executed hg. The warning causes runcmd to panic
since it sees output on stderr.

This patch ignores warnings on stderr about not importing a package.