contrib/wix: fix product name to include version
Without this, Mercurial is registered under the installed programs
on Windows as just 'Mercurial', which is unusual for a program to do
on Windows.

Including the version in the registered product name registers the
software in the Windows control panel of installed programs as, for
example, 'Mercurial 1.5' (for 1.5), or as

  'Mercurial 1.5.1032' for a stable 1.5+32-35893dcfd40c
  'Mercurial 1.5.5080' for a unstable 1.5+80-1ee60e82333c

when using http://bitbucket.org/tortoisehg/thg-winbuild to create
daily builds.

Furthermore, the install UI texts in the built installer database (MSI
file) are more clear on what the user is about to install, as, for example,
the text in the initial installer dialog now reads

  "The Setup Wizard will install Mercurial 1.5 on your computer"

instead of just

  "The Setup Wizard will install Mercurial on your computer"