verify: improve progress descriptions
For GUIs, the progress bar is disconnected visually from the text that
is output by ui.write(), so it's better to be a little bit redundant in
naming the status.  "checking" is redundant, since the user is running
the verify command.  'changesets', 'manifests', etc, tell the user what
is being checked (the entity being counted in the progress bar.