hidden: Add a hiddenrevs attributes to changelog.
This attributes hold the set of all revisions that should be ommited by command
and tools displaying changesets.

This set is given as a hit. Command and tools are responsible to check it in
order to filter they outpur.

Code adding revisions to the set are responsible to the consistency of it's
data.