tests: ignore inotify extension in test-duplicateoptions.py
The inotify extension is only available on linux and setup.py will not install
it on other platforms - but it will of course always be there in the source.

test-duplicateoptions.py tried to load most extensions (including inotify if
available). When the local uninstalled Mercurial was used it would thus always
load the inotify extension and cause a warning on unsupported platforms.

The inotify extension is not relevant for this test, so now we explicitly
ignore it.