discovery: fix prepush documentation
The prepush documentation claim that when we refuse to push, the second element
of the returng tuple is an "outgoing" integer. value should be 0 when no
outgoing changeset and 1 otherwise.  In pratice if there are no outgoing
changeset, "outgoing" value is alway 1 and util.Abort is raised on other error.

the stable branch also include this error so it wasn't introduced by a recent
refactoring.