help: use the first topic name from helptable, not the longest alias
This makes the 'additional help topics' list consistent with the output from
keyword search (for instance subrepo/subrepos).

The sorting by longest name was introduced in 3f6a4f1de43f. There might have
been a good reason for it back then, but now it seems like a better idea to
place the preferred name first in the list in helptable.