merge: infer the "other" changeset when falling back to v1 format
When we have to fallback to the old version of the file, we infer the
"other" from current working directory parent. The same way it is currently done
in the resolve command. This is know to have shortcoming? but we cannot do
better from the data contained in the old file format. This is actually the
motivation to add this new file format.