largefiles: mark lfile as added in lfdirstate when the standin is added
This is an alternative solution to the problem addressed by f72d73937853. This
implementation has the advantage that it doesn't mark clean largefiles as
normallookup. We can thus avoid repeated rehashing of all largefiles when
f72d73937853 is backed out.

This implementation use the existing 'lfmr' actions that 23fe278bde43
introduced for handling another part of the same cases.