windows: adjust doc string and comments of posixfile()
The doc string of osutil.posixfile includes (line 611):

  "On error, this function may raise either a WindowsError or an IOError."

which is most likely correct, but does not fit for this function here anymore,
as we do fold WindowsError to IOError here specifically.

And this function is now a bit more than just an exception-wrapper, as it has
been expanded to additionally sanitize the unloved seek/tell behavior
of Windows.

(Self-disclosure: This patch is entirely untested at the time of its
publication, as I'm currently not using this version myself. I send it
in hopes that it will reduce potential future confusion. CC-ing Matt Harbison)